Move regression test back to its own file
bun-inspect.test.ts has pre-existing 'depth = Infinity' failures on
ASAN/debug builds (shallower stack) that are unrelated to this change.
Keeping the regression test in its own file isolates it from those.
